Fix canary stage timing issues (#2502)
* fix(canary): don't wait an hour before cleanup by default

* fix(canary): Fix up timing of runs to start after deploys complete
